kb-rag-service/requirements.txt
django
djangorestframework
psycopg2-binary
requests
qdrant-client
PyYAML

